refactoring to use docker compose
