+++ README.md
@@ -1,1 +1,1 @@
-# rebalance
+# Rebalancing bot
