+++ pooldb-updater/.dockerignore
@@ -1,0 +1,4 @@
+node_modules
+npm-debug.log
+yarn-error.log
+.DS_Store
