+++ pooldb-updater/src/poolInfo.ts
