+++ pooldb-updater/src/updateAccounts.ts
@@ -1,4 +1,4 @@
-import { LiquidityStateV4, Raydium } from "@raydium-io/raydium-sdk-v2";
+import { Raydium } from "@raydium-io/raydium-sdk-v2";
 import {
   RedisClientType,
   RedisDefaultModules,
@@ -7,7 +7,6 @@ import {
   RedisScripts,
 } from "redis";
 import { getPoolDBData, PoolDBData } from "./poolInfo";
-import { PublicKey } from "@solana/web3.js";
 
 export const updateAccounts = async (
   raydium: Raydium,
@@ -17,9 +16,6 @@ export const updateAccounts = async (
     RedisScripts
   >
 ) => {
-  // rate limit to 3 requests per second
-  //   let lastUpdate = Date.now();
-  //   let counter = 0;
   // eslint-disable-next-line no-constant-condition
   while (true) {
     const nextAmm = await redisClient.lPop("update_queue");
@@ -27,13 +23,6 @@ export const updateAccounts = async (
       await new Promise((resolve) => setTimeout(resolve, 100));
       continue;
     }
-    // while (counter >= 3) {
-    //   if (Date.now() - lastUpdate > 1000) {
-    //     counter = 0;
-    //     lastUpdate = Date.now();
-    //   }
-    // }
-    // counter++;
     const poolDbData = await redisClient.json.get(
       "pools:" + nextAmm.toString()
     );
