+++ pooldb-updater/tsconfig.json
@@ -7,10 +7,6 @@
     "skipLibCheck": true,
     "forceConsistentCasingInFileNames": true,
     "outDir": "./dist",
-    "baseUrl": ".", // Add this line
-    "paths": { // Add this section
-      "@raydium-io/raydium-sdk-v2": ["node_modules/@raydium-io/raydium-sdk-v2"]
-    }
   },
   "include": ["src/**/*"],
   "exclude": ["node_modules"]
