--- src/error.txt
@@ -1,1 +1,0 @@
-Request failed with status code 404
