+++ trader/requirements.txt
@@ -1,0 +1,2 @@
+networkx==3.4.1
+redis==5.0.0
